Added PictureInPicture class
PictureInPicture provides a gst.Bin that puts one video inside another
with half its resolution. The video that will be inside must be
provided in the pad sink_1, while the video that will be outside must be
provided in the pad sink_2. Final resolution and colorspace should be
provided in element instantiation.