Show source configuration interface when the dialog is first opened.
When you open the sources interface and add a new source, the vbox in
the left will not show for the option already selected in the combobox.
If you change the option and change it back, it will appear right.

This adds a new function to set the current factory and pick the
interface and use it both in the init and when changing the option.