Use sltv directory as a package.
This changes our executable file into a python script instead of a shell
script. And it imports sltv_ui as part of the sltv package. This has
required minor changes in the importing of the settings module in the
input UI modules. It has also required the adding of an empty __init__
file for the sltv package.