Split list of sources from UI.
The list of sources should not be in the UI. It also should not be too
much coupled with the GTK TreeModel.

This would require implementing something that looks like a python list
and a TreeModel that uses it. For now, just try to hide the ListStore a
little more.