Edit Source dialog should not be instantiated every time.
Only one dialog is enough. Creating a new dialog every time the add
button is clicked will waste memory without much (if any) gain.

The only difference in behaviour is that the dialog is shown with the
values of the last time it was shown. This may even be considered a
feature, and would not be that difficult to solve.