UI: Play and Stop buttons are now normal buttons.
Instead of using toggle buttons, we use normal buttons for Play and
Stop. To forbid the user of clicking them in the wrong state (click Play
while playing, click Stop while not playing), we set their
sensitiveness. That is, the user cannot click Play while playing or Stop
while not playing.