effects: get ghostpad target pad directly instead of using find_*
Since find_unconnected_pad is deprecated in GStreamer and
find_unlinked_pad is pretty new, we should avoid using them when they
are not strictly necessary.

In this particular case, we know what are the elements in the bin, their
linking order and what pads they should have. So, it's pretty
straightforward to get them as the target pads.