Use get_item to test if a given source exists.
Instead of using knowledge about the internals of the class and try if
there's an already saved source in the config file, we use a given
interface that returns a MediaItem by its name. If it returns None,
there's no MediaItem by that name and we can create a new one.