Connect to SLTV "playing" signal and change buttons state.
We only set the stop button sensitivity to true when we receive a
"playing" signal from SLTV. We still set the sensitivity of the play
button to false both when the play button is pressed and when we receive
the signal. The first one is to forbid the user to press it again when
the playing signal does not come immediately. The second is to disable
the play button when we get the signal even if the user has not pressed
the play button, although it may not happen in any case right now.