Connect to SLTV "stopped" signal and change buttons state.
We only set the play button sensitivity to true when we receive a
"stopped" signal from SLTV. We still set the sensitivity of the stop
button to false both when the stop button is pressed and when we receive
the signal. The first one is to forbid the user to press it again when
the stopped signal does not come immediately. The second is to disable
the stop button when we get the signal even if the user has not pressed
the stop button, which may happen when we receive an EOS.