EditUI set_current_factory set factory to selected one.
When we create a EditUI, we fill its list of factories and the
corresponding combobox. Then, we set the chosen item of combobox to the
first one, and the factory to the first factory in the list.

However, when we sort the combobox liststore, there's no correspondence
anymore in the order of the factory list and the liststore. So, we use a
set_current_factory, that will use the chosen item in the liststore.