UI: split preview checkbox from the core interface
This is the first of a series of splits. Settings widgets like text
overlay, effects, preview checkbox, and others to come will be
independent of the main interface.

This will make it easier to introduce a new interface without
conflicting with other changes. It will also allow us to organize this
interface as we wish later without much hassle.