Added label and separator to audio and output box
Audio and output configurations are the same every time, so the
separator and label can be inside audio_box and output_box. This was not
done for the other boxes (encoding and element specific) because we can
have different encoding boxes and different element specific boxes,
which is not true for audio and output setting boxes.