Adding config for textoverlay
This patch adds font and position configurations for text overlay. It
adds the interface to choose font and position selection and sets those
configurations when apply button is pressed.

The position and font selections are saved in a config file and
retrieved in the next execution of the program.

signed-off: Marcelo Jorge Vieira <metal@holoscopio.com>
signed-off: Luciana Fujii Pontello <luciana@holoscopio.com>