Changed edit.ui to have only general things
Removed from edit.ui specific widgets such as audio, encoding, output
boxes. There is only a container box that is used.

ui/edit.py and ui/edit_output.py had to be changed to work with that,
adding and removing config boxes when needed.

The encoding config is still missing a label and alignment, which will
be fixed when we split encoding from output.