Drop widget from core/preview and put it into UI/core.
The preview area widget was referenced in core through the use of a
constructor parameter. To remove that reference, the Preview object now
emits a signal the UI can watch and link the widget then. The UI uses
the preview member of SLTV object to get the reference to the Preview
object. The signal used gives the GStreamer element and is named after
the XOverlay interface message: "prepare-xwindow-id".