UI/save_button: Return filename in entry instead of dialog.
The FileChooser interface is a little tricky. When we set_filename with
a file that does not exist, it will fail. When we get_filename later, it
will return None or the value that last succeeded.

This will change the value of the filename for a given output to None
(causing a failure when trying to edit the output) or to a different
value (from another file output).

A quick fix is to ignore the value returned by the FileChooserDialog and
use the value in the Entry. However, this will not work for any other
FileChooser we use instead of our custom SaveButton.

Bug-Number: 20
Bug-URL: http://gonod.softwarelivre.org/cgi-bin/bugzilla/show_bug.cgi?id=20