UI/preview: add a widget to connect to a given Preview object
Instead of linking the Gstreamer element into the core, we do this in a
PreviewArea widget. We use a Preview object from the SLTV object to link
both. PreviewArea simply inherits from the GTK+ Drawing Area and
connects to the "prepare-xwindow-id" signal from the Preview object.

This new widget make it easier to create new preview areas that look
into GStreamer elements that implement XOverlay and are embedded into a
Preview object.