sltv: Adding PictureInPicture to SLTV
Each input was divided into two, one stream will be for A video (source
video) and the other the B video (pip video).

A PictureInPicture element was added to the pipeline right after the
split. This class will choose between the sources available. For now,
video A will have a fixed resolution of 320X240, and video B a
resolution of 160X120.

Video B is chosen on the main interface using B button of each source.
It is possible to disable picture-in-picture just pressing the already
selected B button.