Fixing EditSource class using a copy of config dict
When editing a source (or adding a new one) it was affecting the other
sources that had been added during the same execution of the program.
This was happening because we saved a reference for the config
dictionary, and when that dictionary changed when we were adding new
items or editing existing ones, we modified that dictionary.

We are now saving a copy of that dictionary, since each media_item has
to have its own dictionary with its own configuration.

This was already fixed for other media_items other than sources.