UI/overlay: Connect to signals in the end of the constructor.
When we connect to some signals and they are emitted while the
constructor has not yet finished setting the object up, we may end up
executing functions that expect some members in the object that are not
there yet.

To avoid this situation from happening, we only connect to any signal
after all the object is set up.