PIP: Will choose sources as well
PictureInPicture has been redesigned not only to apply one video on
another, but to choose from many videos which are which.

PictureInPicture now creates pads as requested, meaning it can have
inumerous streams linked to it. Pads can be of type A or B (bigger or
smaller). It has always one active pad of each type.