Moving volume setting from settings toolbar
The volume configuration is set by just a button, that is really small
to deserve a settings toolbar. We moved it to the main UI, right next to
play and stop.

XML file created by glade was displaying errors. Since the UI is
basically just a button, we create it by hand.