sltv: Adding all audio inputs to pipeline
Instead of adding only the chosen audio source to the pipeline, we add
all available audio sources, because we want to be able to change the
source during transmition later. To do this, we use the new class
AudioInputBin. All audio sources are connected to input-selector, that
chooses the initially selected audio source.